+++ packages/components/legend/CHANGELOG.md
@@ -1,4 +1,4 @@
-## `@svizzle/legend` v0.2.0 (next)
+## `@svizzle/legend` v0.2.0
 
 - add the field `showTicksExtentOnly` to `flags` prop
 - renamed internal stores to start with an underscore